+++ hugr-core/src/envelope/writer.rs
@@ -1,6 +1,5 @@
 use std::io::Write;
 
-use itertools::Itertools as _;
 use thiserror::Error;
 
 use crate::Hugr;
@@ -82,7 +81,7 @@ fn encode_model_binary<'h>(
 
     // Append extensions for binary model.
     if format == EnvelopeFormat::ModelWithExtensions {
-        serde_json::to_writer(writer, &extensions.iter_all().collect_vec())?;
+        extensions.write_json(writer)?;
     }
 
     Ok(())
@@ -101,7 +100,7 @@ fn encode_model_text<'h>(
 
     // Prepend extensions for text model.
     if format == EnvelopeFormat::SExpressionWithExtensions {
-        serde_json::to_writer(&mut writer, &extensions.iter_all().collect_vec())?;
+        extensions.write_json(&mut writer)?;
     }
 
     let bump = Bump::default();